Глобальный фикс варнингов, xr_3da теперь собирается без варнингов

+++ trunk/xr_3da/GameFont.cpp
@@ -82,7 +82,7 @@ void CGameFont::Initialize		(LPCSTR cShader, LPCSTR cTextureName)
 		uFlags |= fsMultibyte;
 		fHeight = ini->r_float( "mb_symbol_coords" , "height" );
 		
-		fXStep = ceil( fHeight / 2.0f );
+		fXStep = static_cast<float>(ceil(static_cast<double>(fHeight / 2.0f)));
 
 		for ( u32 i=0 ; i < nNumChars ; i++ ) {
 			sprintf_s( buf ,sizeof(buf), "%05d" , i );
